plugin-ext: validate path when unpacking archives
Fix zip-slip by validating where a given file will be unpacked. If the
expected path is outside of the destination folder: log a warning and
ignore the file.

This commit includes an archive that will trigger the exploit by writing
a file to `/tmp/slipped.txt`. This comes from
https://github.com/kevva/decompress/issues/71#issue-502469706

Fixes https://github.com/eclipse-theia/theia/issues/7319

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>